GCC: Silence Warnings

+++ src/maid-x11.c
@@ -306,11 +306,11 @@ static XImage *ReadBMP(Display *dpy, char *Name)
 		return (NULL);
 	}
 
-	for (y = 0; y < infoheader.biHeight; y++)
+    for (y = 0; y < (int)infoheader.biHeight; y++)
 	{
 		int y2 = infoheader.biHeight - y - 1;
 
-		for (x = 0; x < infoheader.biWidth; x++)
+        for (x = 0; x < (int)infoheader.biWidth; x++)
 		{
 			int ch = getc(f);
